third_party/blink/web_tests/platform/generic/media/media-fragments/TC0070-expected.txt


Title: TC0070
Fragment: 't=1e-1'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
